+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 - added: Re-enable Piratechain on iOS
 - added: Battery Saver warning message for Android
 - added: Buy conversion tracking for Moonpay
+- added: Buy conversion tracking for Paybis
+- added: Support for return.edge.app deeplinks
 - added: Error tracking for failure to report conversions to referral server
 - changed: Remove whitespaces from custom token contract address input
 - changed: Use unique ENV configs for thorchain and thorchainda swap plugins
 - fixed: Correctly tag `tokenApproval` `actionType` in  `getTxActionDisplayInfo`
+- fixed: URI encoding in Paybis return URIs
 - fixed: AddressTile2 touchable area states
 - fixed: Cases where it was possible to create duplicate custom tokens
 - fixed: Clear previous swap errors when new amounts are entered or swap assets are changed in `SwapCreateScene`